Improves Mui combobox selection by label
Enhances the `selectMuiByLabel` function to better identify and interact with Mui combobox elements.

It now prioritizes elements directly associated with the label via the 'for' attribute.
Also, dispatches keydown events to open the combobox and enter events to select and dispatches 'input' event to simulate text typing.

This addresses issues where the correct combobox was not always selected, or the selection was not properly triggered.